feat(tui): render the source view's diff overlay in split view
Wires ADR 0049's split rendering into draw_source_screen: Body::Split
via render_scrollable_pane, falling back to the existing unified
overlay below MIN_SPLIT_VIEW_WIDTH or when split_source_rows returns
None (drift). v/V (InputKey::ToggleSplitView) now also flips the
shared diff_view_mode from Screen::Source, matching its already-global
help text and status-line hint (both corrected to match).

Splits source_screen's TestBackend tests into a source_screen_tests/
directory by topic (mirroring diff_pane_tests/), adding a split_view
module alongside the existing error_handling/highlighting/overlay
groups.
